# swagger/ChatStack.yaml
openapi: 3.0.1
info:
  title: ChatStack API
  version: "1.0"
  description: Chatting Platform
...
servers:
  - url: /chat-stack/api/v1
security:
  - bearerAuth: []

# ===========================================================================
# COMPONENTS
# ===========================================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      required:
        - id
        - email
        - fullname
        - display_name
        - country
        - email_verified
        - status
        - created_at
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
          description: Must be a valid email address
        fullname:
          type: string
        display_name:
          type: string
        email_verified:
          type: boolean
        avatar_url:
          type: string
          nullable: true
        status:
          type: string
          enum: [online, offline]
        timezone:
          type: string
          description: IANA timezone
          example: Europe/Paris
        last_seen_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required:
        - reason
      properties:
        reason:
          type: string
          description: >-
            HTTP error or free text such as Database Not Available,
            Invalid Credential, Mandatory, Invalid Format, Invalid Length,
            ForbiddenValue, Invalid Data, etc.
        code:
          type: string
          description: Server / SQL error ID
        description:
          type: string
          description: Human-readable error description including data and value

    # -------------------------------------------------------------------------
    # Auth Request Bodies
    # -------------------------------------------------------------------------
    SignupRequest:
      type: object
      required:
        - email
        - password
        - fullname
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          example: user@example.com
          description: Must be a valid email address
        password:
          type: string
          writeOnly: true
          format: password
          minLength: 12
          maxLength: 30
          pattern: "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&#]).*$"
          description: >
            Must be at least 12 characters and contain uppercase, lowercase,
            a digit, and a special character (@$!%*?&#).
        fullname:
          type: string
          minLength: 2
          maxLength: 255
          example: John Doe
        display_name:
          type: string
          minLength: 3
          maxLength: 50
          pattern: "^[a-zA-Z0-9._-]+$"
          description: Unique public username
          example: john_doe

    # Password login — email + password only
    PasswordLoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
          description: Must be a valid email address
        password:
          type: string
          format: password
          minLength: 12
          maxLength: 30
          pattern: "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&#]).*$"
          writeOnly: true
          description: >
            Must be at least 12 characters and contain uppercase, lowercase,
            a digit, and a special character (@$!%*?&#).

    # OAuth2 login — provider ID token exchange
    OAuth2LoginRequest:
      type: object
      required:
        - id_token
      properties:
        id_token:
          type: string
          description: >
            ID token obtained from the OAuth2 provider (e.g. the credential
            returned by Google Sign-In). The server validates it and either
            logs in the linked account or creates a new one.
        device_name:
          type: string
          description: Optional human-readable device label
          example: "Firefox on Ubuntu"

    # -------------------------------------------------------------------------
    # Auth Response Bodies
    # -------------------------------------------------------------------------

    # Returned on login / token refresh — refresh token travels via HttpOnly cookie
    AuthResponse:
      type: object
      required:
        - access_token
        - user
      properties:
        access_token:
          type: string
          description: Short-lived JWT (e.g. 15 min)
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        user:
          $ref: '#/components/schemas/User'

    # Returned on signup — same shape, separate name for clarity
    SignupResponse:
      type: object
      required:
        - access_token
        - user
      properties:
        access_token:
          type: string
          description: Short-lived JWT. Email must be verified before full access.
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        user:
          $ref: '#/components/schemas/User'
    RefreshResponse:
      type: object
      required:
        - access_token
      properties:
        access_token:
          type: string
          description: New short-lived JWT. Old refresh token is invalidated (rotation).
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
    AuthResendVerificationRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          example: test@example.com

    # -------------------------------------------------------------------------
    # User Update Bodies
    # -------------------------------------------------------------------------
...
paths:
...
  /auth/signup:
    post:
      operationId: authSignup
      tags:
        - Authentication Flow
      summary: Create a new account with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: >
            Account created. A refresh token is set as an HttpOnly cookie.
            The user must verify their email before accessing protected resources.
          headers:
            Set-Cookie:
              description: >
                HttpOnly, Secure, SameSite=Strict refresh token cookie.
                Scoped to path /chat-stack/api/v1/auth/refresh-token.
              schema:
                type: string
                example: "refreshToken=<token>; HttpOnly; Secure; SameSite=Strict; Path=/chat-stack/api/v1/auth/refresh-token"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignupResponse'
        '400':
          description: ValidationUtils error (invalid email, weak password, etc.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      operationId: authLogin
      tags:
        - Authentication Flow
      summary: Login with email and password
      description: >
        Authenticates with credentials. On success, issues a short-lived access token
        in the response body and rotates a refresh token via an HttpOnly cookie.
        Each successful login creates a new session row (multi-device support).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordLoginRequest'
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              description: >
                HttpOnly, Secure, SameSite=Strict refresh token cookie.
                Scoped to path /chat-stack/api/v1/auth/refresh-token.
              schema:
                type: string
                example: "refreshToken=<token>; HttpOnly; Secure; SameSite=Strict; Path=/chat-stack/api/v1/auth/refresh-token"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Wrong email or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Email not yet verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many login attempts — rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/oauth2/{provider}:
    post:
      operationId: authOAuth2Login
      tags:
        - Authentication Flow
      summary: Login or register via an OAuth2 provider (e.g. Google)
      description: >
        Exchanges the provider's ID token for a ChatStack session.
        If the provider account is not yet linked to any user, a new account is
        created automatically (no separate signup step needed).
        If the email already exists under a password account, a 409 is returned
        with a link-account hint.
      security: []
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google, github]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OAuth2LoginRequest'
      responses:
        '200':
          description: Existing account — login successful
          headers:
            Set-Cookie:
              description: HttpOnly refresh token cookie (same as password login)
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '201':
          description: New account created via OAuth2 provider
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignupResponse'
        '400':
          description: Invalid or expired id_token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: >
            Email already registered via password. Prompt user to link accounts
            or login with their password instead.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      operationId: authLogout
      tags:
        - Authentication Flow
      summary: Logout — revoke one session, all sessions, or a specific device
      description: >
        If `session_id` is omitted and `all_devices` is false (the default),
        only the current session (identified by the refresh token cookie) is revoked.
        Use `session_id` to log out a specific device. Use `all_devices: true`
        to revoke every active session for this user.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                session_id:
                  type: string
                  format: uuid
                  description: >
                    Target a specific session to revoke.
                    Must belong to the authenticated user.
                all_devices:
                  type: boolean
                  default: false
                  description: If true, revokes all sessions regardless of session_id.
      responses:
        '204':
          description: Session(s) revoked successfully
          headers:
            Set-Cookie:
              description: Clears the refreshToken cookie
              schema:
                type: string
                example: "refreshToken=; HttpOnly; Secure; SameSite=Strict; Path=/chat-stack/api/v1/auth/refresh-token; Max-Age=0"
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: session_id does not belong to the authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: session_id not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh-token:
    post:
      operationId: authRefreshToken
      tags:
        - Authentication Flow
      summary: Rotate refresh token and issue a new access token
      description: >
        Reads the refresh token from the HttpOnly cookie.
        Validates the token, issues a new access token, rotates the refresh
        token (old one is immediately invalidated — token family rotation).
        If the presented token has already been rotated (reuse detected),
        the entire token family is revoked to contain a potential breach.
      security: []   # authenticated via HttpOnly cookie, not Bearer
      responses:
        '200':
          description: New access token issued; new refresh token set in cookie
          headers:
            Set-Cookie:
              description: Rotated HttpOnly refresh token cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        '401':
          description: Missing, invalid, or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: >
            Refresh token already revoked or reuse detected —
            all sessions in the token family have been invalidated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/verify-email:
    post:
      operationId: authVerifyEmail
      tags:
        - Authentication Flow
      summary: Verify user email with token from the verification link
      security: []
      parameters:
        - name: token
          in: query
          required: true
          description: >
            The verification token sent by email. Can also be sent in the request body.
          schema:
            type: string
      responses:
        '204':
          description: Email verified successfully
        '400':
          description: Token is missing, malformed, or already used
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '410':
          description: Verification token has expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/resend-verification:
    post:
      operationId: authResendVerification
      tags:
        - Authentication Flow
      summary: Resend the email verification link
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  description: Must be a valid email address
      responses:
        '204':
          description: Verification email resent (or silently ignored if already verified)
        '400':
          description: Email already verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many resend requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/forgot-password:
    post:
      operationId: authForgotPassword
      tags:
        - Authentication Flow
      summary: Request a password reset email
      description: >
        Always returns 204 regardless of whether the email is registered,
        to avoid user enumeration.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthResendVerificationRequest'
      responses:
        '204':
          description: Reset email sent (if account exists)
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/reset-password:
    post:
      operationId: authResetPassword
      tags:
        - Authentication Flow
      summary: Reset password using the token from the reset email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - new_password
              properties:
                new_password:
                  type: string
                  format: password
                  writeOnly: true
                  minLength: 12
                  maxLength: 30
                  pattern: "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&#]).*$"
                  description: >
                    Must be at least 12 characters and contain uppercase, lowercase,
                    a digit, and a special character (@$!%*?&#).
      responses:
        '204':
          description: Password reset successfully
        '400':
          description: Invalid password format or token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '410':
          description: Reset token expired or already used
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/change-password:
    post:
      operationId: authChangePassword
      tags:
        - Authentication Flow
      summary: Change password while authenticated (knows current password)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - current_password
                - new_password
              properties:
                email:
                  type: string
                  format: email
                  description: Must be a valid email address
                current_password:
                  type: string
                  format: password
                  minLength: 12
                  maxLength: 30
                  pattern: "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&#]).*$"
                  writeOnly: true
                  description: >
                    Must be at least 12 characters and contain uppercase, lowercase,
                    a digit, and a special character (@$!%*?&#).
                new_password:
                  type: string
                  format: password
                  writeOnly: true
                  minLength: 12
                  maxLength: 30
                  pattern: "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&#]).*$"
                  description: >
                    Must be at least 12 characters and contain uppercase, lowercase,
                    a digit, and a special character (@$!%*?&#).
      responses:
        '204':
          description: Password changed successfully
        '400':
          description: Current password is wrong or new password fails validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ===========================================================================
  # SESSION MANAGEMENT
  # ===========================================================================
...